Del opp i flere komponenter
